+++ src/main/java/com/newsportal/controller/UserController.java
@@ -5,17 +5,18 @@ import com.newsportal.dto.UserDTO;
 import com.newsportal.dto.UserUpdateDTO;
 import com.newsportal.model.User;
 import com.newsportal.service.UserService;
+import jakarta.servlet.http.HttpSession;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
-
-import java.util.List;
 
 
 @RestController
 @RequestMapping("/api/users")
 public class UserController {
-
+    @Autowired
+    private HttpSession session;
     @Autowired
     private UserService userService;
 
@@ -25,6 +26,16 @@ public class UserController {
                 .map(user -> new UserDTO(user.getUsername(), user.getEmail())) // Convert User to UserDTO
                 .map(ResponseEntity::ok)
                 .orElseGet(() -> ResponseEntity.notFound().build());
+    }
+
+    @GetMapping("/profile")
+    public ResponseEntity<?> getCurrentUserProfile() {
+        User currentUser = (User) session.getAttribute("user");
+        if (currentUser == null) {
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body("User is not logged in.");
+        }
+        UserDTO userDTO = new UserDTO(currentUser.getUsername(), currentUser.getEmail());
+        return ResponseEntity.ok(userDTO);
     }
 
     @PostMapping("/create")
@@ -41,29 +52,49 @@ public class UserController {
 
     @DeleteMapping("delete/{id}")
     public ResponseEntity<?> deleteNews(@PathVariable Long id) {
+        if (!userService.isCurrentUserAdmin()) {
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).body("Access denied.");
+        }
         userService.deleteUser(id);
         return ResponseEntity.ok("User item deleted successfully");
     }
 
     @GetMapping("/admins")
-    public ResponseEntity<List<User>> getAllAdminUsers() {
+    public ResponseEntity<?> getAllAdminUsers() {
+      if (!userService.isCurrentUserAdmin()) {
+
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).body("Access denied.");
+        }
         return ResponseEntity.ok(userService.getAllAdminUsers());
     }
 
     @GetMapping("/users")
-    public ResponseEntity<List<User>> getAllUsers() {
+    public ResponseEntity<?> getAllUsers() {
+        if (!userService.isCurrentUserAdmin()) {
+
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).body("Access denied.");
+        }
         return ResponseEntity.ok(userService.getAllUsers());
     }
 
     @GetMapping("/waiting")
-    public ResponseEntity<List<User>> getAllWaitingUsers() {
+    public ResponseEntity<?> getAllWaitingUsers() {
+        if (!userService.isCurrentUserAdmin()) {
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).body("Access denied.");
+        }
         return ResponseEntity.ok(userService.getAllWaitingUsers());
     }
 
 
     @PutMapping("/update")
     public ResponseEntity<String> updateUSer(@RequestBody UserUpdateDTO userUpdateDTO) {
+        if (!userService.isCurrentUserAdmin()) {
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).body("Access denied.");
+        }
         userService.updateUser(userUpdateDTO);
         return ResponseEntity.ok("User updated successfully");
     }
+
+
+
 }
